refactor(search): extract find_path_occupied into primitives/path.rs
`find_path_occupied` was defined as a `pub(crate)` helper inside
`drivers/entropy.rs`, but two non-entropy consumers
(`generators/greedy.rs` and `placement/nohome.rs`) had to reach into
the entropy driver for it. That coupling was one of the §4 architectural
findings in the 2026-05-31 review ("entropy.rs is two modules in one":
a search engine plus a utility provider for unrelated callers).

Move the function and its two unit tests to a new `primitives/path.rs`
module, promote visibility to `pub`, and update the three callers:

- `drivers/entropy.rs` (the original definition site, plus its in-
  function call inside `sequential_fallback`),
- `generators/greedy.rs`,
- `placement/nohome.rs`.

Added a third test (`find_path_occupied_source_equals_target`) covering
the `from == to` early-return, which the original tests didn't pin.

Net effect on `drivers/entropy.rs`: -67 lines (function body + ribbon
comments + two tests). The driver is now solely the entropy-search
engine; its utility role is gone.

Verified:

- 284 search-crate unit tests pass (was 283; +1 from the new
  source-equals-target case).
- 60-case `temp_regression` integration suite passes bit-for-bit.
- `cargo clippy -p bloqade-lanes-search --all-targets -- -D warnings`
  is clean.
