+++ crates/bloqade-lanes-search/src/primitives/mod.rs
@@ -11,3 +11,4 @@ pub mod distance;
 pub mod graph;
 pub mod lane_index;
 pub(crate) mod ordering;
+pub mod path;
